+++ tests/integration/resource_options/policy-pack-python/__main__.py
@@ -16,6 +16,10 @@ def create_urn(type_: str, name: str) -> str:
     return f"urn:pulumi:{get_stack()}::{get_project()}::{type_}::{name}"
 
 
+def create_stack_urn() -> str:
+    return create_urn("pulumi:pulumi:Stack", f"{get_project()}-{get_stack()}")
+
+
 def validate_resource(args, report_violation):
     validate(args)
 
@@ -60,6 +64,7 @@ def validate(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     else:
         raise AssertionError(f"Unexpected resource of type: '{t}'.")
@@ -74,6 +79,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "protect":
         options_equal(PolicyResourceOptions(
@@ -83,6 +89,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "ignoreChanges":
         options_equal(PolicyResourceOptions(
@@ -92,6 +99,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "deleteBeforeReplaceNotSet":
         options_equal(PolicyResourceOptions(
@@ -101,6 +109,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "deleteBeforeReplaceTrue":
         options_equal(PolicyResourceOptions(
@@ -110,6 +119,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "deleteBeforeReplaceFalse":
         options_equal(PolicyResourceOptions(
@@ -119,6 +129,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "aliased":
         options_equal(PolicyResourceOptions(
@@ -133,6 +144,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "timeouts":
         options_equal(PolicyResourceOptions(
@@ -142,6 +154,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(60, 120, 180),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "timeouts-create":
         options_equal(PolicyResourceOptions(
@@ -151,6 +164,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(240, 0, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "timeouts-update":
         options_equal(PolicyResourceOptions(
@@ -160,6 +174,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 300, 0),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "timeouts-delete":
         options_equal(PolicyResourceOptions(
@@ -169,6 +184,7 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=[],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 360),
+            parent=create_stack_urn(),
         ), r.opts)
     elif r.name == "secrets":
         options_equal(PolicyResourceOptions(
@@ -178,6 +194,27 @@ def validate_dynamic_resource(r):
             aliases=[],
             additional_secret_outputs=["foo"],
             custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
+        ), r.opts)
+    elif r.name == "parent-res":
+        options_equal(PolicyResourceOptions(
+            protect=False,
+            ignore_changes=[],
+            delete_before_replace=None,
+            aliases=[],
+            additional_secret_outputs=["foo"],
+            custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_stack_urn(),
+        ), r.opts)
+    elif r.name == "child-res":
+        options_equal(PolicyResourceOptions(
+            protect=False,
+            ignore_changes=[],
+            delete_before_replace=None,
+            aliases=[],
+            additional_secret_outputs=["foo"],
+            custom_timeouts=PolicyCustomTimeouts(0, 0, 0),
+            parent=create_urn("pulumi-nodejs:dynamic:Resource", "parent-res"),
         ), r.opts)
     else:
         raise AssertionError(f"Unexpected resource with name: '{r.name}'.")